+++ util/genrest/protomodelcreator.go
@@ -16,13 +16,20 @@ package genrest
 
 import (
 	"fmt"
+	"io/ioutil"
+	"os"
 
+	"github.com/ghodss/yaml"
 	"github.com/golang/protobuf/protoc-gen-go/descriptor"
 	"github.com/googleapis/gapic-showcase/util/genrest/internal/pbinfo"
 	"github.com/googleapis/gapic-showcase/util/genrest/protomodel"
 	"google.golang.org/genproto/googleapis/api/annotations"
+	"google.golang.org/genproto/googleapis/api/serviceconfig"
+	"google.golang.org/genproto/googleapis/longrunning"
 	"google.golang.org/protobuf/compiler/protogen"
+	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/reflect/protodesc"
 	"google.golang.org/protobuf/types/descriptorpb"
 )
 
@@ -50,24 +57,43 @@ func NewProtoModel(plugin *protogen.Plugin) (*protomodel.Model, error) {
 		for _, svc := range protoFile.GetService() {
 			serviceModel := protoModel.AddService(NewService(protoPackage, svc))
 			for _, method := range svc.GetMethod() {
-				options := method.GetOptions()
-				if options == nil {
-					continue
-				}
+				addBindingsForMethod(protoModel, serviceModel, method)
+			}
+		}
+	}
 
-				eHTTP /*, err*/ := proto.GetExtension(method.GetOptions(), annotations.E_Http)
-				http := eHTTP.(*annotations.HttpRule)
-				rules := []*annotations.HttpRule{http}
-				rules = append(rules, http.GetAdditionalBindings()...)
-				for idxRule, oneRule := range rules {
-					protoModel.AccumulateError(NewServiceBinding(serviceModel, method, oneRule, idxRule))
-				}
-
+	serviceConfig, err := GetServiceConfig()
+	if err != nil {
+		return nil, err
+	}
+	mixins := collectMixins(serviceConfig)
+	for _, mixinFile := range mixins {
+		protoPackage := *mixinFile.file.Package
+		for _, mixinService := range mixinFile.services {
+			svc := mixinService.service
+			serviceModel := protoModel.AddService(NewService(protoPackage, svc))
+			for _, method := range mixinService.methods {
+				addBindingsForMethod(protoModel, serviceModel, method)
 			}
 		}
 	}
 
 	return protoModel, protoModel.Error()
+}
+
+func addBindingsForMethod(protoModel *protomodel.Model, serviceModel *protomodel.Service, method *descriptor.MethodDescriptorProto) {
+	options := method.GetOptions()
+	if options == nil {
+		return
+	}
+
+	eHTTP /*, err*/ := proto.GetExtension(method.GetOptions(), annotations.E_Http)
+	http := eHTTP.(*annotations.HttpRule)
+	rules := []*annotations.HttpRule{http}
+	rules = append(rules, http.GetAdditionalBindings()...)
+	for idxRule, oneRule := range rules {
+		protoModel.AccumulateError(NewServiceBinding(serviceModel, method, oneRule, idxRule))
+	}
 }
 
 ////////////////////////////////////////
@@ -140,4 +166,113 @@ func NewRESTRequestPattern(rule *annotations.HttpRule) (*protomodel.RESTRequestP
 		return nil, fmt.Errorf("unhandled pattern: %#x", pattern)
 	}
 	return binding, nil
+}
+
+////////////////////////////////////////
+// Mixins
+
+// Note: much of the code below is a modified copy of the mixin code in gapic-generator-go.
+
+// GetServiceConfig reads and returns the Showcase service config file.
+func GetServiceConfig() (*serviceconfig.Service, error) {
+	// TODO: Consider getting this from the plugin options. On the
+	// other hand, there's only one copy of this file, so maybe
+	// hard-coding this location isn't terrible.
+	serviceConfigPath := "schema/google/showcase/v1beta1/showcase_v1beta1.yaml"
+
+	y, err := ioutil.ReadFile(serviceConfigPath)
+	if err != nil {
+		cwd, _ := os.Getwd()
+		return nil, fmt.Errorf("error reading service config %q (cwd==%q): %v", serviceConfigPath, cwd, err)
+	}
+
+	j, err := yaml.YAMLToJSON(y)
+	if err != nil {
+		return nil, fmt.Errorf("error converting YAML to JSON: %v", err)
+	}
+
+	serviceConfig := &serviceconfig.Service{}
+	if err := (protojson.UnmarshalOptions{DiscardUnknown: true}).Unmarshal(j, serviceConfig); err != nil {
+		return nil, fmt.Errorf("error unmarshaling service config: %v", err)
+	}
+
+	// An API Service Config will always have a `name` so if it is not populated,
+	// it's an invalid config.
+	if serviceConfig.GetName() == "" {
+		return nil, fmt.Errorf("invalid API service config file %q", serviceConfigPath)
+	}
+	return serviceConfig, nil
+}
+
+// Mixins is the collection of files containing methods to be mixed in.
+type Mixins []*MixinFile
+
+// MixinFile describes a single file containins methods to be mixed in.
+type MixinFile struct {
+	file     *descriptor.FileDescriptorProto
+	services []*MixinService
+}
+
+// MixinService describes a single service containing methods to be filled in
+type MixinService struct {
+	service *descriptor.ServiceDescriptorProto
+	methods []*descriptor.MethodDescriptorProto
+}
+
+// indexedRules keys HTTP rules by their selectors
+type indexedRules map[string]*annotations.HttpRule
+
+// collectMixins collects the configured mixin APIs from the service config and
+// returns the appropriately configured mixin `MethodDescriptorProto`s to generate for each.
+func collectMixins(serviceConfig *serviceconfig.Service) Mixins {
+	mixinRules := indexedRules{}
+	for _, rule := range serviceConfig.GetHttp().GetRules() {
+		mixinRules[rule.GetSelector()] = rule
+	}
+	mixins := Mixins{}
+	for _, api := range serviceConfig.GetApis() {
+		mixins = append(mixins, getMixinsForAPI(mixinRules, api.GetName())...)
+	}
+	return mixins
+}
+
+// getMixinsForAPI returns the appropriately configured mixin `MethodDescriptorProto`s
+// corresponding to the `mixinRules` that refer to `api`. The method descriptors are taken from the
+// package-global `mixinDescriptors`.
+func getMixinsForAPI(mixinRules indexedRules, api string) Mixins {
+	files := Mixins{}
+	for _, file := range mixinDescriptors[api] {
+		fileToAdd := &MixinFile{
+			file: file,
+		}
+		files = append(files, fileToAdd)
+		for _, service := range file.GetService() {
+			serviceToAdd := &MixinService{
+				service: service,
+			}
+			fileToAdd.services = append(fileToAdd.services, serviceToAdd)
+			for _, method := range service.GetMethod() {
+				fqn := fmt.Sprintf("%s.%s.%s", file.GetPackage(), service.GetName(), method.GetName())
+
+				if rule := mixinRules[fqn]; rule != nil {
+					proto.SetExtension(method.Options, annotations.E_Http, rule)
+					serviceToAdd.methods = append(serviceToAdd.methods, method)
+				}
+			}
+		}
+	}
+	return files
+}
+
+// mixinDescriptors maps fully qualified proto service names of mixins implemented by Showcase to a
+// list of FileDescriptors containing the definitions needed for that service.
+var mixinDescriptors map[string][]*descriptor.FileDescriptorProto
+
+func init() {
+	// TODO(noahdietz): Figure out how to add Locations and IAMPolicy to this list.
+	mixinDescriptors = map[string][]*descriptor.FileDescriptorProto{
+		"google.longrunning.Operations": {
+			protodesc.ToFileDescriptorProto(longrunning.File_google_longrunning_operations_proto),
+		},
+	}
 }